// src/models/IRequiredReports.ts
export interface IRequiredReports {
    isStatementSalesRequired: boolean;
    isStatementIncomesRequired: boolean;
    isCustomerFlightsRequired: boolean;
    isCustomerBuhDocsRequired: boolean;
    isCustomerAirCompaniesRequired: boolean;
    isCustomerTicketOrdersRequired: boolean;
    isCustomerAdditionalService: boolean;
}